// src/app/models/event.ts
export interface Event {
        idEvent: number;
        idSoccerXML: string;
        idAPIfootball: any;
        strEvent: string;
        strEventAlternate: string;
        strFilename: string;
        strSport: string;
        idLeague: number;
        strLeague: string;
        strSeason: string;
        strDescriptionEN: any;
        strHomeTeam: string;
        strAwayTeam: string;
        intHomeScore: number;
        intRound: number;
        intAwayScore: number;
        intSpectators: number;
        strHomeGoalDetails: string;
        strHomeRedCards: string;
        strHomeYellowCards: string;
        strHomeLineupGoalkeeper: string;
        strHomeLineupDefense: string;
        strHomeLineupMidfield: string;
        strHomeLineupForward: string;
        strHomeLineupSubstitutes: string;
        strHomeFormation: any;
        strAwayRedCards: string;
        strAwayYellowCards: string;
        strAwayGoalDetails: string;
        strAwayLineupGoalkeeper: string;
        strAwayLineupDefense: string;
        strAwayLineupMidfield: string;
        strAwayLineupForward: string;
        strAwayLineupSubstitutes: string;
        strAwayFormation: any;
        intHomeShots: number;
        intAwayShots: number;
        dateEvent: string;
        dateEventLocal: any;
        strDate: string;
        strTime: string;
        strTimeLocal: string;
        strTVStation: any;
        idHomeTeam: number;
        idAwayTeam: number;
        strResult: any;
        strCircuit: any;
        strCountry: any;
        strCity: any;
        strPoster: any;
        strFanart: any;
        strThumb: any;
        strBanner: any;
        strMap: any;
        strTweet1: any;
        strTweet2: any;
        strTweet3: any;
        strVideo: any;
        strLocked: string;
}